# Add argument for enabling xFormers optimizations