This change allows users to take advantage of the speed and memory optimizations provided by [xFormers](https://github.com/facebookresearch/xformers). Since this package is currently not published with binaries for all platforms (like Windows) I didn't add it to `requirements.txt` but rather made it an optional install for now.

`enable_xformers_memory_efficient_attention()` gives a big boost in speed while also reducing GPU memory usage significantly. A 512x512 test I did cut the processing time almost in half: 

Without xFormers:
```
2022-11-28 17:29:56.669 | INFO     | lama_cleaner.model.base:_pad_forward:54 - final forward pad size: (512, 512, 3)
100%|██████████████████████████████████████████████████████████████████████████████████| 51/51 [00:12<00:00,  4.16it/s]
2022-11-28 17:30:12.514 | INFO     | lama_cleaner.server:process:164 - process time: 15846.312522888184ms
```
With xFormers:
```
2022-11-28 17:30:48.092 | INFO     | lama_cleaner.model.base:_pad_forward:54 - final forward pad size: (512, 512, 3)
100%|██████████████████████████████████████████████████████████████████████████████████| 51/51 [00:05<00:00,  8.70it/s]
2022-11-28 17:30:57.286 | INFO     | lama_cleaner.server:process:164 - process time: 9196.5172290802ms
```